test: 添加生产源码行数守卫
