+++ packages/weapp-tailwindcss/src/bundlers/shared/generator-css/source-resolver/configuration.ts
@@ -1,6 +1,6 @@
 import type { IStyleHandlerOptions } from '@weapp-tailwindcss/postcss/types'
 import type { GeneratorResolvedSource } from './metadata'
-import type { GeneratorSourceRuntimeState, GeneratorSourceSelectionOptions, SourceStyleMatchOptions, TailwindV4CssSource, TailwindV4CssSourceRef, TailwindV4SourceOptions } from './types'
+import type { GeneratorSourceRuntimeState, SourceStyleMatchOptions, TailwindV4CssSource, TailwindV4CssSourceRef, TailwindV4SourceOptions } from './types'
 import type { TailwindResolvedSource } from '@/generator'
 import { existsSync, readFileSync } from 'node:fs'
 import path from 'node:path'
@@ -12,10 +12,9 @@ import { normalizeConfigDirective } from '../config-directive'
 import { hasTailwindRootDirectives, hasTailwindSourceDirectives, resolveCssEntrySource } from '../directives'
 import { hasTailwindGeneratedCss, hasTailwindGeneratedCssMarkers } from '../markers'
 import { resolveExistingConfigPath } from './config'
-import { normalizeCssSourceForCompare, scoreTailwindV4CssSourceFileMatch } from './matching'
+import { normalizeCssSourceForCompare } from './matching'
 import { withGeneratorSourceMetadata } from './metadata'
-import { resolvePostcssFromOption, resolvePostcssSourceFile } from './postcss-source'
-import { resolveSingleTailwindV4CssSource, resolveTailwindV4CssEntryIncludesPreflight, resolveTailwindV4CssSourceEntries } from './single-source'
+import { resolvePostcssFromOption } from './postcss-source'
 
 export function createCssEntrySources(cssEntries: string[] | undefined) {
   return cssEntries
@@ -155,63 +154,6 @@ export function shouldResolveSourceSideCssEntry(rawSource: string) {
     || hasTailwindGeneratedCssMarkers(rawSource)
 }
 
-export async function resolveMatchingTailwindV4CssEntry(
-  rawSource: string,
-  file: string,
-  sourceOptions: ReturnType<typeof resolveTailwindV4SourceOptionsFromRuntime>,
-) {
-  const cssEntries = sourceOptions.cssEntries
-  if (!cssEntries?.length) {
-    return undefined
-  }
-
-  const normalizedFile = path.resolve(file.replace(/[?#].*$/, ''))
-  const pathMatchedEntries = cssEntries.filter(cssEntry =>
-    path.resolve(cssEntry.replace(/[?#].*$/, '')) === normalizedFile,
-  )
-  if (pathMatchedEntries.length === 1) {
-    return resolveTailwindV4CssEntrySource(pathMatchedEntries[0]!, sourceOptions, {
-      includesPreflight: await resolveTailwindV4CssEntryIncludesPreflight(pathMatchedEntries[0]!),
-      index: cssEntries.indexOf(pathMatchedEntries[0]!),
-    })
-  }
-
-  const normalizedRawSource = normalizeCssSourceForCompare(rawSource)
-  const matches = cssEntries
-    .map((cssEntry) => {
-      if (!existsSync(cssEntry)) {
-        return undefined
-      }
-      try {
-        const entrySource = readFileSync(cssEntry, 'utf8')
-        const pathScore = scoreTailwindV4CssSourceFileMatch(file, cssEntry, sourceOptions)
-        if (normalizeCssSourceForCompare(entrySource) === normalizedRawSource) {
-          return {
-            cssEntry,
-            score: 1000000 + pathScore,
-          }
-        }
-        return undefined
-      }
-      catch {
-        return undefined
-      }
-    })
-    .filter((match): match is { cssEntry: string, score: number } => Boolean(match))
-    .sort((a, b) => b.score - a.score)
-  const bestScore = matches[0]?.score
-  const matchingEntry = bestScore && matches.filter(match => match.score === bestScore).length === 1
-    ? matches[0]?.cssEntry
-    : undefined
-  if (!matchingEntry) {
-    return undefined
-  }
-  return resolveTailwindV4CssEntrySource(matchingEntry, sourceOptions, {
-    includesPreflight: await resolveTailwindV4CssEntryIncludesPreflight(matchingEntry),
-    index: cssEntries.indexOf(matchingEntry),
-  })
-}
-
 export function normalizeTailwindV4CssSourceConfig(
   cssSource: TailwindV4CssSource,
   sourceBase: string,
@@ -307,61 +249,6 @@ export function normalizeTailwindV4CssSourceConfigs(sourceOptions: TailwindV4Sou
         cssSources,
       }
     : sourceOptions
-}
-
-export async function resolveMatchingTailwindV4CssSource(
-  rawSource: string,
-  file: string,
-  cssHandlerOptions: IStyleHandlerOptions,
-  sourceOptions: TailwindV4SourceOptions,
-  selectionOptions?: GeneratorSourceSelectionOptions | undefined,
-) {
-  const cssSources = sourceOptions.cssSources
-  if (!cssSources?.length) {
-    return undefined
-  }
-
-  const normalizedRawSource = normalizeCssSourceForCompare(rawSource)
-  const sourceFile = resolvePostcssSourceFile(cssHandlerOptions)
-  const matches = cssSources
-    .map((cssSource, index) => {
-      if (typeof cssSource.css !== 'string' || cssSource.css.length === 0) {
-        return undefined
-      }
-      if (sourceFile && typeof cssSource.file === 'string' && path.resolve(sourceFile) === path.resolve(cssSource.file)) {
-        return {
-          cssSource,
-          index,
-          score: 1000000,
-        }
-      }
-      if (normalizeCssSourceForCompare(cssSource.css) === normalizedRawSource) {
-        const pathScore = typeof cssSource.file === 'string'
-          ? scoreTailwindV4CssSourceFileMatch(file, cssSource.file, sourceOptions)
-          : 0
-        return {
-          cssSource,
-          index,
-          score: 1 + pathScore,
-        }
-      }
-      return undefined
-    })
-    .filter((match): match is { cssSource: TailwindV4CssSource, index: number, score: number } => Boolean(match))
-    .sort((a, b) => b.score - a.score || a.index - b.index)
-  const bestScore = matches[0]?.score
-  const matchingSource = bestScore && matches.filter(match => match.score === bestScore).length === 1
-    ? matches[0]?.cssSource
-    : undefined
-  if (!matchingSource) {
-    return undefined
-  }
-  return resolveSingleTailwindV4CssSource(matchingSource, sourceOptions, {
-    includesPreflight: await resolveTailwindV4CssSourceEntries(matchingSource, sourceOptions).then(resolved => resolved?.includesPreflight),
-    index: matches[0]?.index,
-    matched: true,
-    primary: isSameTailwindV4CssSource(matchingSource, selectionOptions?.cssSources?.[0]),
-  })
 }
 
 export function tryResolveTailwindV4SourceOptions(
